refactor: fold the site Camera type into data-renderers/types
common/camera.ts existed for the zoom and pan copies that Camera2D
replaced. All that was left was the Camera type, whose six consumers are
layers.ts and the data-renderers beside it - five of which already
imported from ./types, so the move collapses into existing imports.
